image_pro/backend/requirements.txt
Flask
flask-cors
pillow
piexif